feat: add withCollapsableHeader hoc
